Se crean endpoints de movimientos
Se añaden endpoints para el registro de movimientos y tambien para la generacion de informes por rango de fechas

+++ apiRest/src/main/java/com/controller/CuentaController.java
@@ -25,7 +25,7 @@ public class CuentaController {
     
 
     @GetMapping("/{numeroCuenta}")
-    public ResponseEntity<Cuenta> obtenerCuentaPorNumero(@PathVariable Long numeroCuenta) {
+    public ResponseEntity<Cuenta> obtenerCuentaPorNumero(@PathVariable int numeroCuenta) {
         Cuenta cuenta = cuentaService.buscarPorNumero(numeroCuenta);
         return ResponseEntity.ok(cuenta);
     }
